# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for Teralux App.",
        "title": "Teralux API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/tuya/devices/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves details of a specific device by its ID. Response includes last_commands field containing the last control commands sent to the device.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "02. Devices"
                ],
                "summary": "Get Device by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.TuyaDeviceResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    }
                }
            }
        },
...
        "/api/tuya/devices/{id}/sensor": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves sensor data (temperature, humidity, etc.) for a specific device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "04. Device Sensor"
                ],
                "summary": "Get Sensor Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.SensorDataDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.SensorDataDTO": {
            "type": "object",
            "properties": {
                "battery_percentage": {
                    "type": "integer"
                },
                "humidity": {
                    "type": "integer"
                },
                "status_text": {
                    "type": "string"
                },
                "temp_unit": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                }
            }
        },
        "dtos.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "dtos.TuyaDeviceDTO": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "collections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.TuyaDeviceDTO"
                    }
                },
                "create_time": {
                    "type": "integer"
                },
                "custom_name": {
                    "type": "string"
                },
                "gateway_id": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "local_key": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "online": {
                    "type": "boolean"
                },
                "product_name": {
                    "type": "string"
                },
                "remote_category": {
                    "type": "string"
                },
                "remote_id": {
                    "type": "string"
                },
                "remote_product_name": {
                    "type": "string"
                },
                "status": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.TuyaDeviceStatusDTO"
                    }
                },
                "update_time": {
                    "type": "integer"
                }
            }
        },
        "dtos.TuyaDeviceResponseDTO": {
            "type": "object",
            "properties": {
                "device": {
                    "$ref": "#/definitions/dtos.TuyaDeviceDTO"
                }
            }
        },
        "dtos.TuyaDeviceStatusDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}